requirements.txt
python-dotenv~=1.0.0
chatgpt4pcg~=1.1.2